+++ docs/changelog.rst
@@ -13,6 +13,10 @@ Development
     - run_in_transaction context manager relies on Pymongo coreAPI, it will retry automatically in case of `UnknownTransactionCommitResult` but not `TransientTransactionError` exceptions
     - Using .count() in a transaction will always use Collection.count_document (as estimated_document_count is not supported in transactions)
 - Fix use of $geoNear or $collStats in aggregate #2493
+- Further to the deprecation warning, remove ability to use an unpacked list to `Queryset.aggregate(*pipeline)`, a plain list must be provided instead `Queryset.aggregate(pipeline)`, as it's closer to pymongo interface
+- Further to the deprecation warning, remove `full_response` from `QuerySet.modify` as it wasn't supported with Pymongo 3+
+- Fixed stacklevel of many warnings (to point places emitting the warning more accurately)
+- Add support for collation/hint/comment to delete/update and aggregate #2842
 
 Changes in 0.29.0
 =================